requirement.txt
streamlit==1.31.0
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4
sqlalchemy==2.0.25
python-docx==1.1.0
bcrypt==4.1.2
python-dateutil==2.8.2
Pillow==10.2.0
